refactor: simplified copy button components
